GO-7143 Re-enable desktop memory-growth detector, write proper snapshots
Extract the snapshot writer into core/debug/debugsnapshot so callers
outside core/application can produce snapshot archives without
depending on the Service. The application service becomes a thin
wrapper that supplies wallet identity + stat.json via
debugsnapshot.Meta.

Re-enable the memory-growth detector in core/debug/profiler on desktop
builds (gomobile stays a no-op). Previously it logged a base64-encoded
heap profile at Error level into zap, expecting the gelf sink to ship
it to Graylog — with gelf gone the blob just rotated out of
anytype.log. It now calls debugsnapshot.Save and writes a proper
snapshot zip into common/profiles/, so DebugExportReport can ship it
the same way as client-triggered snapshots.

Detector behaviour unchanged: 1GiB system-memory threshold, 1.5x
growth factor, max 3 snapshots per run, 1-minute sampling interval.

No new API surface; the build-tag split (profiler_desktop.go vs
profiler_mobile.go) already restricts this to desktop binaries.
