+++ core/debug/profiler/service.go
@@ -36,9 +36,8 @@ func (s *service) Name() (name string) {
 }
 
 func (s *service) Run(ctx context.Context) (err error) {
-	// todo: temporarily disable until GO-7143
-	// go s.run()
-
+	// Desktop-only: run() is a no-op on gomobile (see profiler_mobile.go).
+	go s.run()
 	return nil
 }
